requirements.txt
loguru==0.5.3
numpy==1.18.5
pandas==1.1.1
Flask_WTF==0.14.2
dash_core_components==0.30.0rc1
plotly==4.3.0
Flask_Bcrypt==0.7.1
dash_html_components==1.0.3
Flask==1.0.3
WTForms==2.2.1
dash==1.10.0
Flask_Admin==1.5.7
Flask_SQLAlchemy==2.4.1
Flask_Login==0.4.1
Pillow==8.0.1
python_dateutil==2.8.1
secrets==1.0.2
